docs(readme): document how to use snippets
Persisting the changes I had to make to get everything (including
snippets) working with LuaSnip, the default snippets provider in
LazyVim.

Neovim 0.10 has native snippet support, so once that has a stable
release we probably want to document that as the default.

packer.nvim has been deprecated, so move its documentation section to
the bottom.